Update map with current location (#5)
* Updates Location on Map

Updates the Location on Google Maps, by using ViewModel and MutableLiveData. Moved the logic of the RxObservables to the ISSClient file.

* Fixed Bug, Added Back Button

Fixed the bug that if the android back button was pressed from MapFragment, the MainActivity would not update its Lat and Long. Also added a Floating Action Button to add an in app route to the MainActivity.